ML

Description

This module defines components for modular construction of small domain specific languages.

Declaring Languages

type family DeclareLanguage features prim :: * -> * where ...

Declare a new language with the given primitives, and some additional features.

Equations

DeclareLanguage (f ': more) prim = f (DeclareLanguage more prim)

DeclareLanguage '[] prim = prim

class (forall m. Language m => Language (f m)) => Feature f

Promotes statements in a language to an extended language.

Minimal complete definition

lift

Primitives

data Pure a

A language with no-built in primitives.

data IO a

Language Features

data Val x t m a

Extends language m with an immutable variable x of type t.

data Mut x t m a

Extends language m with an mutable variable x of type t.

data Col x t m a

Extends language m with a write-only variable x of type t.

data Throws t m a

Extends language m with support for exception of type t.

data Backtracks m a

Extends language m with support for backtracking.

Writing Programs

Generic Language Combinators

replicateM_ :: Language m => Int -> m a -> m ()

Repeat a statement some number of times, ignoring its result.

replicateM :: Language m => Int -> m a -> m [a]

Repeat a statement some number of times, collecting the results.

forM_ :: (Foldable col, Language m) => col a -> (a -> m b) -> m ()

Repeat a statement for each member of a collection, ignoring the statemnt's result.

forM :: (Traversable col, Language m) => col a -> (a -> m b) -> m (col b)

Repeat a statement for each member of a collection, and produce a new collection containing the answers.

forever :: Language m => m a -> m b

Repeat the given statement forever, ignore its result.

when :: Language m => Bool -> m () -> m ()

Do the statement when the condition is true.

unless :: Language m => Bool -> m () -> m ()

Do the statement unless the condition is true (i.e., when it is false).

Introducing Specific Effects

class (Language p, Language m) => HasPrim p m | m -> p where

Language m supports the primitives defined by language p.

Methods

prim :: p a -> m a

Promote a primitive statement to language m

class Language m => HasVal x t m | m x -> t where

Language m supports an immutable variable x of type t.

Methods

readVal :: x -> m t

Get the value of variable x.

class Language m => HasMut x t m | m x -> t where

Language m supports a mutable variable x of type t.

Methods

getMut :: x -> m t

Get the value of variable x.

setMut :: x -> t -> m ()

Set the value of variable x.

class Language m => HasCol x t m | m x -> t where

Language m supports a write-only variable x of type t.

Methods

appendTo :: x -> t -> m ()

Append the given value to variable x.

class Language m => CanThrow t m where

Language m supports throwing exceptions of type t.

Methods

throw :: t -> m a

Throw an exception of type t.

class Language m => CanBacktrack m where

Language m supports backtracking.

Methods

backtrack :: m a

Backtrack, abandoning the current execution path.

orElse :: m a -> m a -> m a

Introduce a choice point---if the first computation backtracks, then proceed with the second.

Compiling Programs

val :: Language m => (x := t) -> Val x t m a -> m a

Given a value for x, compile a program that relies on immutable variable x, to one that does not use x explicitly.

data x := t

Just a convenient notatino for initializing variables.

Constructors

x := t

mut :: Language m => (x := t) -> Mut x t m a -> m (a, t)

Give an initial value for x, compile a program that relies on a mutable variable x, to one that does not use x explicitly. The resulting program returns the final value of x in addition to its result.

collector :: Language m => Col x t m a -> m (a, [t])

Compile a program with a collector x to one that does not collect explicitly. The new program returns the collected values, in addition to the original program's result.

throws :: Language m => Throws t m a -> m (Except t a)

Compile a program that may throw an exception of type t to one that does not do so explicitly. The resulting program will wither produce a sucessful result, or an exception of one was thrown. See Except.

data Except t a

The result of a program that may throw exceptions.

Constructors

Exception t
The program threw an exception.

Success a
The program completed successfully.

backtracks :: Language m => Maybe Int -> Backtracks m a -> m [a]

Compile a program that may backtrack, to one that does not do so explictly. The resulting program may return zero or more answers. The input parameter optionally specifies an upper bound on the nubmer of required answers.

noEffects :: Pure a -> a

A program with no effects can be compiled to a value.

withIO :: IO a -> IO a

Primitive operations are not compiled any further.

class Language m => Run m where

Compile a program all the way to the primitive language.

Associated Types

type ExeResult m a

Methods

run :: m a -> ExeResult m a

Nested Effects

class HasVal x t m => LetVal x t m | x m -> t where

Language m supports temporary modifications to an immutable variable for the duration of the given program block.

Methods

letVal :: (x := t) -> m a -> m a

class HasCol x t m => CanCollect x t m where

Language m supports collecting the output of a nested program block. The resulting statement produces no additional output to x.

Methods

collect :: x -> m a -> m (a, [t])

class CanThrow t m => CanCatch t m where

Language m supports runing a local computation and "catching" exceptions. The answer of type Except indicates if an exception occured. The resulting program is guarnateed to not throw an excpetion of the given type.

Methods

try :: m a -> m (Except t a)

class CanBacktrack m => CanSearch m where

Language m supports finding the answers of a local backtracking block.

Methods

findUpTo :: Maybe Int -> m a -> m [a]

The input parameter optionally imposes an upper bound on the number of answers to return.

Misc

type Language = Monad

A language is another word for monad.